requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.11.7
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.0
dill==0.3.5.1
discord.py==2.0.0
frozenlist==1.3.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
multidict==6.0.2
platformdirs==2.5.2
pylint==2.14.5
tomli==2.0.1
tomlkit==0.11.4
wrapt==1.14.1
yarl==1.8.1
